pyproject.toml
[tool.poetry]
name = "external_market_information"
version = "0.1.0"
description = "Collect external information for specified companies, like news and stock prices."
authors = ["marco_schwarz <marco.schwarz@onelogic.de>"]

[tool.poetry.dependencies]
python = "^3.8"
logging = "^0.4.9"
urllib3 = "^1.26.3"
pandas = "^1.2.3"
numpy = "^1.20.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
